Fix clippy warnings: collapse nested if statements and simplify code
- arch.rs: change filter_map to map (always returns Some), use char array for find
- deb.rs: collapse nested if let statements
- gpg.rs: collapse multiple nested if statements in key verification and listing
- fedora.rs: collapse nested if statements in XML parsing
- selector.rs: collapse filter conditions
- commands/install.rs, commands/mod.rs: collapse nested if statements

All code now passes `cargo clippy -- -D warnings`

+++ src/commands/install.rs
@@ -219,11 +219,11 @@ pub fn cmd_install(
         let mut changeset = conary::db::models::Changeset::new(changeset_desc);
         let changeset_id = changeset.insert(tx)?;
 
-        if let Some(old_trove) = old_trove_to_upgrade.as_ref() {
-            if let Some(old_id) = old_trove.id {
-                info!("Removing old version {} before upgrade", old_trove.version);
-                conary::db::models::Trove::delete(tx, old_id)?;
-            }
+        if let Some(old_trove) = old_trove_to_upgrade.as_ref()
+            && let Some(old_id) = old_trove.id
+        {
+            info!("Removing old version {} before upgrade", old_trove.version);
+            conary::db::models::Trove::delete(tx, old_id)?;
         }
 
         let mut trove = rpm.to_trove();
